app/requirements.txt
schedule==1.2.2
openai==1.75.0
PyPDF2==3.0.1
Jinja2==3.1.6
weasyprint==65.1
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
torch==2.7.0
openai-whisper==20240930
